DSP 0.22WIP1
Enhanced NEC CPU, added 6 new arcade, added i2c eprom

+++ src/cpu/z80/z80_sp.pas
@@ -101,7 +101,7 @@ var
  instruccion,temp:byte;
  posicion:parejas;
  ban_temp:band_z80;
- pcontador:byte;
+ pcontador:integer;
  irq_temp:boolean;
  cantidad_t:word;
 begin
